add database initial
