+++ tables.sql
@@ -1,0 +1,14 @@
+/*
+-- Query: SHOW tables from bd_activa_app
+-- Date: 2022-11-19 13:10
+*/
+INSERT INTO `` (`Tables_in_bd_activa_app`) VALUES ('activa_work');
+INSERT INTO `` (`Tables_in_bd_activa_app`) VALUES ('activa_work_student_rel');
+INSERT INTO `` (`Tables_in_bd_activa_app`) VALUES ('activa_work_tech_skill_rel');
+INSERT INTO `` (`Tables_in_bd_activa_app`) VALUES ('job');
+INSERT INTO `` (`Tables_in_bd_activa_app`) VALUES ('reward');
+INSERT INTO `` (`Tables_in_bd_activa_app`) VALUES ('soft_skill');
+INSERT INTO `` (`Tables_in_bd_activa_app`) VALUES ('student');
+INSERT INTO `` (`Tables_in_bd_activa_app`) VALUES ('tech_skill');
+INSERT INTO `` (`Tables_in_bd_activa_app`) VALUES ('training');
+INSERT INTO `` (`Tables_in_bd_activa_app`) VALUES ('user');
