--- users.sql
@@ -1,7 +1,0 @@
-/*
--- Query: SELECT * FROM bd_activa_app.user
-LIMIT 0, 1000
-
--- Date: 2022-11-18 23:37
-*/
-INSERT INTO `` (`id`,`email`,`password`,`role`) VALUES (1,'oscar.villaplana@xarxatecactiva.com','123456789','estudiante');
